test: refactor and extract util for asserting on bytecode version
