+++ github-workflows-kt/src/test/kotlin/io/github/typesafegithub/workflows/JavaBytecodeVersionTest.kt
@@ -1,29 +1,15 @@
 package io.github.typesafegithub.workflows
 
 import io.github.typesafegithub.workflows.domain.Workflow
+import io.github.typesafegithub.workflows.testutils.JdkVersionToBytecodeVersion.JDK_11
+import io.github.typesafegithub.workflows.testutils.shouldHaveBytecodeVersion
 import io.kotest.core.spec.style.FunSpec
-import io.kotest.matchers.shouldBe
-import java.io.DataInputStream
 
 class JavaBytecodeVersionTest :
     FunSpec(
         {
             test("library is built with desired Java bytecode version") {
-                Workflow::class.java.classLoader
-                    .getResourceAsStream(
-                        "io/github/typesafegithub/workflows/domain/Workflow.class",
-                    )!!
-                    .use { inputStream ->
-                        val dataInputStream = DataInputStream(inputStream)
-                        require(dataInputStream.readInt() == 0xCAFEBABE.toInt()) { "Invalid class header" }
-                        val minor = dataInputStream.readUnsignedShort()
-                        val major = dataInputStream.readUnsignedShort()
-
-                        // Corresponds to JDK 11
-                        // See https://javaalmanac.io/bytecode/versions/
-                        major shouldBe 55
-                        minor shouldBe 0
-                    }
+                Workflow::class shouldHaveBytecodeVersion JDK_11
             }
         },
     )
