+++ test-utils/src/main/kotlin/io/github/typesafegithub/workflows/testutils/JdkVersionToBytecodeVersion.kt
@@ -1,0 +1,8 @@
+package io.github.typesafegithub.workflows.testutils
+
+/**
+ * See https://javaalmanac.io/bytecode/versions/
+ */
+object JdkVersionToBytecodeVersion {
+    const val JDK_11 = "55.0"
+}
